pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sre-incident-env"
version = "1.0.0"
description = "SRE Incident Response Environment for OpenEnv — AI agents triage production incidents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "websockets>=13.0",
    "requests>=2.31.0",
]

[project.scripts]
server = "uvicorn:main"

[tool.setuptools.packages.find]
include = ["sre_incident_env*"]

[tool.openenv]
name = "sre_incident_env"
module = "sre_incident_env.server.app"
app = "app"
environment = "sre_incident_env.server.environment:SREIncidentEnvironment"
action = "sre_incident_env.models:SREIncidentAction"
observation = "sre_incident_env.models:SREIncidentObservation"
state = "sre_incident_env.models:SREIncidentState"
